apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
numpy==1.26.4